[CMLG-007] INSTANCE-SEARCHIING Issues: Trying to make the code more understandable Solution: Rename the file, change the code in App.js, delete the useless code in index.css. Risk-Area: -- Reviewed by: --.
